refactor: examples
